HW 3.4

+++ src/main/java/ru/hogwarts/schoolspring/impl/FacultyServiceImp.java
@@ -2,19 +2,25 @@ package ru.hogwarts.schoolspring.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.hogwarts.schoolspring.exeptions.FacultyNotFoundException;
 import ru.hogwarts.schoolspring.model.Faculty;
+import ru.hogwarts.schoolspring.model.Student;
 import ru.hogwarts.schoolspring.repositories.FacultyRepository;
+import ru.hogwarts.schoolspring.repositories.StudentRepository;
 import ru.hogwarts.schoolspring.service.FacultyService;
 import java.util.Collection;
+import java.util.Optional;
 
 @Service
 public class FacultyServiceImp implements FacultyService {
 
     @Autowired
     private final FacultyRepository facultyRepository;
+    private final StudentRepository studentRepository;
 
-    public FacultyServiceImp(FacultyRepository facultyRepository) {
+    public FacultyServiceImp(FacultyRepository facultyRepository, StudentRepository studentRepository) {
         this.facultyRepository = facultyRepository;
+        this.studentRepository = studentRepository;
     }
 
     @Override
@@ -24,17 +30,30 @@ public class FacultyServiceImp implements FacultyService {
 
     @Override
     public Faculty getFaculty(long id) {
-        return facultyRepository.findById(id).orElseThrow();
+        Optional<Faculty> faculty = facultyRepository.findById(id);
+        if (faculty.isPresent()) {
+            return facultyRepository.findById(id).orElseThrow();
+        }
+        throw new FacultyNotFoundException();
     }
 
     @Override
     public Faculty editFaculty(Faculty faculty) {
-        return facultyRepository.save(faculty);
+        Optional<Faculty> facul = facultyRepository.findById(faculty.getId());
+        if (facul.isPresent()) {
+            return facultyRepository.save(faculty);
+        }
+        throw new FacultyNotFoundException();
     }
 
     @Override
-    public void deleteFaculty(long id) {
-        facultyRepository.deleteById(id);
+    public Faculty deleteFaculty(long id) {
+        Optional<Faculty> faculty = facultyRepository.findById(id);
+        if (faculty.isPresent()) {
+            facultyRepository.deleteById(id);
+            return faculty.get();
+        }
+        throw new FacultyNotFoundException();
     }
 
     @Override
@@ -45,5 +64,15 @@ public class FacultyServiceImp implements FacultyService {
     @Override
     public Collection<Faculty> getAllFacultyByColor(String color) {
         return facultyRepository.getAllFacultyByColor(color);
+    }
+
+    @Override
+    public Collection<Faculty> getAllFacultyByColorOrName(String colorOrName) {
+        return facultyRepository.getAllFacultyByColorOrName(colorOrName, colorOrName);
+    }
+
+    @Override
+    public Collection<Student> findStudentsByFacultyId(long id) {
+        return studentRepository.findByFaculty_Id(id);
     }
 }
